Move services and repositories to database

+++ src/main/java/edu/kpi/backend/service/UserService.java
@@ -1,7 +1,9 @@
 package edu.kpi.backend.service;
 
 import edu.kpi.backend.dto.CreateUserDTO;
+import edu.kpi.backend.entity.Account;
 import edu.kpi.backend.entity.User;
+import edu.kpi.backend.repository.AccountRepository;
 import edu.kpi.backend.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,25 +15,40 @@ import java.util.UUID;
 @Service
 public class UserService {
     private final UserRepository userRepository;
+    private final AccountRepository accountRepository;
 
     @Autowired
-    public UserService(UserRepository userRepository) {
+    public UserService(UserRepository userRepository, AccountRepository accountRepository) {
         this.userRepository = userRepository;
+        this.accountRepository = accountRepository;
     }
 
     public List<User> getAllUsers() {
-        return this.userRepository.getAll();
+        return this.userRepository.findAll();
     }
 
     public Optional<User> getUserById(UUID id) {
-        return this.userRepository.getById(id);
+        return this.userRepository.findById(id);
     }
 
     public User createUser(CreateUserDTO createUserDTO) {
-        return this.userRepository.create(createUserDTO.getName());
+        Account account = new Account();
+        User user = new User(createUserDTO.getName(), account);
+
+        this.accountRepository.save(account);
+        this.userRepository.save(user);
+
+        return user;
     }
 
     public Optional<User> deleteUserById(UUID id) {
-        return this.userRepository.deleteById(id);
+        Optional<User> user = this.userRepository.findById(id);
+
+        if (user.isPresent()) {
+            this.accountRepository.deleteById(user.get().getAccount().getId());
+            this.userRepository.deleteById(id);
+        }
+
+        return user;
     }
 }
